Move spinner package under internal/pkg/term/ directory
