+++ internal/pkg/term/spinner/charset.go
