+++ internal/pkg/term/spinner/charset_windows.go
