+++ internal/pkg/term/spinner/mocks/mock_spinner.go
