+++ internal/pkg/term/spinner/spinner.go
